# Conflicts:
#	sources/cmake/debugger.cmake
#	sources/plugin_cpp_api/CMakeLists.txt
#	sources/plugins/Rfxcom/Transceiver.cpp
#	sources/plugins/YadomsPluginsCommon.cmake
#	sources/scriptInterpreters/YadomsScriptInterpertersCommon.cmake
#	sources/script_cpp_api/CMakeLists.txt
#	sources/server/database/common/adapters/HugeDataVectorAdapter.hpp
#	sources/server/web/rest/service/Acquisition.cpp
#	sources/shared/CMakeLists.txt
#	sources/www/js/objects/configuration/combo-section-parameter-handler.js
#	sources/www/js/objects/page-manager.js
#	tests/unit/sources/CMakeLists.txt
#	tests/unit/sources/plugins/dev-FakePlugin/TestFakePluginConfiguration.cpp
#	tests/unit/sources/shared/event/TestEventTimer.cpp

+++ sources/plugins/Sigfox/SigfoxConfiguration.h
@@ -25,7 +25,7 @@ public:
    double getRssiMax() const override;
    double getTensionMin() const override;
    double getTensionMax() const override;
-   int getSocketPort() const override;
+   unsigned short getSocketPort() const override;
    // [END] ISigfoxConfiguration implementation
 
 private:
